requirements.txt
python-dotenv
openai-whisper
pyannote.audio
ffmpeg-python
torch==2.3.0 --extra-index-url https://download.pytorch.org/whl/cpu
torchaudio==2.3.0 --extra-index-url https://download.pytorch.org/whl/cpu
numpy==1.26.4
huggingface_hub
transformers
spacy
streamlit